# docs/swagger.yaml
definitions:
...
  dto.Action:
    properties:
      body:
        description: 身體部位(1:全身/2:核心/3:手臂/4:背部/5:臀部/6:腿部/7:肩膀/8:胸部)
        example: 8
        type: integer
      category:
        description: 分類(1:重量訓練/2:有氧/3:HIIT/4:徒手訓練/5:其他)
        example: 1
        type: integer
      cover:
        description: 封面
        example: 32as1d5f13e4.png
        type: string
      equipment:
        description: 器材(1:無需任何器材/2:啞鈴/3:槓鈴/4:固定式器材/5:彈力繩/6:壺鈴/7:訓練椅/8:瑜珈墊/9:其他)
        example: 1
        type: integer
      favorite:
        description: 是否收藏(0:否/1:是)
        example: 1
        type: integer
      id:
        description: 動作id
        example: 1
        type: integer
      intro:
        description: 動作介紹
        example: 槓鈴胸推是很多人在健身房都會訓練的動作，是胸大肌強化最常見的訓練動作
        type: string
      name:
        description: 動作名稱
        example: 槓鈴臥推
        type: string
      source:
        description: 動作來源(1:系統動作/2:教練自創動作)
        example: 2
        type: integer
      type:
        description: 紀錄類型(1:重訓/2:時間長度/3:次數/4:次數與時間/5:有氧)
        example: 1
        type: integer
      video:
        description: 動作影片
        example: 11d547we1d4f8e.mp4
        type: string
    type: object
...
  dto.WorkoutAudio:
    properties:
      audio:
        description: 語音檔案名
        example: e6d2131w5q.mp3
        type: string
    type: object
...
  dto.WorkoutSet:
    properties:
      action:
        $ref: '#/definitions/dto.Action'
        description: 動作
      auto_next:
        description: 自動下一組(Y:是/N:否)
        example: "N"
        type: string
      distance:
        description: 距離(公里)
        example: 0
        type: number
      duration:
        description: 時長(秒)
        example: 30
        type: integer
      id:
        description: 訓練組id
        example: 10
        type: integer
      incline:
        description: 坡度
        example: 0
        type: number
      progress_audio:
        description: 進行中語音
        example: 1d2w3e51d3w.mp3
        type: string
      remark:
        description: 備註
        example: ˋ注意姿勢
        type: string
      reps:
        description: 次數
        example: 0
        type: integer
      start_audio:
        description: 前導語音
        example: 1d2w3e51d3w.mp3
        type: string
      type:
        description: 動作類別(1:動作/2:休息)
        example: 2
        type: integer
      weight:
        description: 重量(公斤)
        example: 0
        type: number
    type: object
  dto.WorkoutSetID:
    properties:
      workout_set_id:
        description: 訓練組id
        example: 10
        type: integer
    type: object
...
  model.Data:
    type: object
  model.ErrorResult:
    properties:
      code:
        description: 錯誤碼
        example: 9000
        type: integer
      data:
        $ref: '#/definitions/model.Data'
        description: 回傳資料
      msg:
        description: 錯誤訊息
        example: system error!
        type: string
    type: object
...
  model.SuccessResult:
    properties:
      code:
        description: 狀態碼
        example: 0
        type: integer
      data:
        $ref: '#/definitions/model.Data'
        description: 回傳資料
      msg:
        description: 成功訊息
        example: success!
        type: string
    type: object
...
  validator.DuplicateWorkoutSetBody:
    properties:
      duplicate_count:
        description: 複製個數
        example: 1
        type: integer
    required:
    - duplicate_count
    type: object
...
  validator.UpdateWorkoutSetBody:
    properties:
      auto_next:
        description: 自動下一組(Y:是/N:否)
        example: "Y"
        type: string
      distance:
        description: 距離(0.01~999.99公里)
        example: 20.25
        type: number
      duration:
        description: 時長(1~38439秒)
        example: 60
        type: integer
      incline:
        description: 坡度(0.01~999.99)
        example: 15.5
        type: number
      remark:
        description: 備註(1~20字元)
        example: 注意呼吸
        type: string
      reps:
        description: 次數(1~999次)
        example: 10
        type: integer
      weight:
        description: 重量(0.01~999.99公斤)
        example: 10.25
        type: number
    type: object
...
info:
  contact: {}
  description: 健身平台 api
  title: fitness api
paths:
...
  /v1/workout_set/{workout_set_id}:
    delete:
      consumes:
      - application/json
      description: 刪除訓練組
      parameters:
      - description: 訓練組id
        in: path
        name: workout_set_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 刪除成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.WorkoutSetID'
              type: object
        "400":
          description: 刪除失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 刪除訓練組
      tags:
      - WorkoutSet_v1
    patch:
      consumes:
      - application/json
      description: 修改訓練組
      parameters:
      - description: 訓練組id
        in: path
        name: workout_set_id
        required: true
        type: integer
      - description: 輸入參數
        in: body
        name: json_body
        required: true
        schema:
          $ref: '#/definitions/validator.UpdateWorkoutSetBody'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.WorkoutSet'
              type: object
        "400":
          description: 更新失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 修改訓練組
      tags:
      - WorkoutSet_v1
  /v1/workout_set/{workout_set_id}/duplicate:
    post:
      consumes:
      - application/json
      description: 複製訓練組
      parameters:
      - description: 訓練組id
        in: path
        name: workout_set_id
        required: true
        type: integer
      - description: 輸入參數
        in: body
        name: json_body
        required: true
        schema:
          $ref: '#/definitions/validator.DuplicateWorkoutSetBody'
      produces:
      - application/json
      responses:
        "200":
          description: 複製成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.WorkoutSet'
                  type: array
              type: object
        "400":
          description: 複製失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 複製訓練組
      tags:
      - WorkoutSet_v1
  /v1/workout_set/{workout_set_id}/progress_audio:
    delete:
      consumes:
      - application/json
      description: 刪除訓練組進行中語音
      parameters:
      - description: 訓練組id
        in: path
        name: workout_set_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 刪除成功!
          schema:
            $ref: '#/definitions/model.SuccessResult'
        "400":
          description: 刪除失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 刪除訓練組進行中語音
      tags:
      - WorkoutSet_v1
    post:
      consumes:
      - multipart/form-data
      description: '下載訓練組進行中語音 : https://www.fitopia-hub.tk/api/v1/resource/workout_set/progress_audio/{語音檔案名}'
      parameters:
      - description: 訓練組id
        in: path
        name: workout_set_id
        required: true
        type: integer
      - description: 進行中語音
        in: formData
        name: progress_audio
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.WorkoutAudio'
              type: object
        "400":
          description: 失敗!
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 上傳訓練組進行中語音
      tags:
      - WorkoutSet_v1
  /v1/workout_set/{workout_set_id}/start_audio:
    delete:
      consumes:
      - application/json
      description: 刪除訓練組前導語音
      parameters:
      - description: 訓練組id
        in: path
        name: workout_set_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 刪除成功!
          schema:
            $ref: '#/definitions/model.SuccessResult'
        "400":
          description: 刪除失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 刪除訓練組前導語音
      tags:
      - WorkoutSet_v1
    post:
      consumes:
      - multipart/form-data
      description: '下載訓練組前導語音 : https://www.fitopia-hub.tk/api/v1/resource/workout_set/start_audio/{語音檔案名}'
      parameters:
      - description: 訓練組id
        in: path
        name: workout_set_id
        required: true
        type: integer
      - description: 前導語音
        in: formData
        name: start_audio
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.WorkoutAudio'
              type: object
        "400":
          description: 失敗!
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 上傳訓練組前導語音
      tags:
      - WorkoutSet_v1
securityDefinitions:
  fitness_token:
    in: header
    name: Token
    type: apiKey
swagger: "2.0"